Added more signal loss response options

+++ src/backend/mathphysics/unit.rs
@@ -7,7 +7,7 @@ pub type MeterPerMillisecond = f32;
 pub type MeterPerSecond = f32;
 pub type Megahertz = u32;
 // TODO add proper type in SI
-pub type PowerUnit = u16;
+pub type PowerUnit = u32;
 
 
 // Const for conversion from km / (s * MHz) to m / (s * Hz).
